+++ dev_build/tool/run_ci_menu.dart
@@ -1,0 +1,8 @@
+import 'package:dev_build/menu/menu_io.dart';
+import 'package:dev_build/menu/menu_run_ci.dart';
+
+Future main(List<String> args) async {
+  mainMenuConsole(args, () {
+    runCiMenu('.');
+  });
+}